feat: move email quote extractor
