+++ app/javascript/dashboard/helper/emailQuoteExtractor.js
