+++ app/javascript/dashboard/helper/specs/emailQuoteExtractor.spec.js
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest';
-import { EmailQuoteExtractor } from '../removeReply.js';
+import { EmailQuoteExtractor } from '../emailQuoteExtractor.js';
 
 const SAMPLE_EMAIL_HTML = `
 <p>method</p>
